requirements.txt
asgiref==3.2.3
blinker==1.4
Django==3.0.2
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
mongoengine==0.19.1
pymongo==3.10.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
